Update README.md
The Dice Game project is a simple and easy two-player game where the players roll a virtual dice by refreshing the website. The game randomly generates a number on the dice each time the website is refreshed, and the player with the highest number on their dice wins the game. If both players have the same number, then it's a tie.

The project is implemented using HTML, CSS, and JavaScript. The game is designed to be played in a web browser and requires no additional software or installation.

The game is a fun way to spend some time with friends and family and can be a great icebreaker for parties and gatherings. It is a simple and easy project suitable for beginners who want to learn web development. The project is open source and contributions are always welcome.