Update README.md
Dice Game

This is a simple and easy two-player game, where the winner is the one who gets more dice when the website is refreshed.

How to Play

Open the website in your web browser.
Two players should be ready to play the game.
Refresh the website to roll the dice.
The player with the highest number on their dice wins the game.
If both players have the same number, then it's a tie.
Installation
Clone the repository or download the zip file and extract it.
bash
Copy code
git clone https://github.com/KamilBatu/Dice.git
Open the index.html file in your web browser.
Contributing
Contributions are always welcome. If you want to contribute to this project, please follow the steps below:

Fork the repository.
Create a new branch.
Make changes and test them.
Submit a pull request.